test: cover key manager with basic tests

+++ src/contracts/implementations/key-registries/KeyRegistry.sol
@@ -9,14 +9,14 @@ contract KeyRegistry is KeyManager, IKeyRegistry {
     /**
      * @inheritdoc IKeyRegistry
      */
-    function getOperatorsLengthAt(uint48 timestamp, bytes memory hint) public view virtual returns (uint256) {
+    function getOperatorsLengthAt(uint48 timestamp, bytes memory hint) public view virtual returns (uint208) {
         return _getKeysOperatorsLengthAt(timestamp, hint);
     }
 
     /**
      * @inheritdoc IKeyRegistry
      */
-    function getOperatorsLength() public view virtual returns (uint256) {
+    function getOperatorsLength() public view virtual returns (uint208) {
         return _getKeysOperatorsLength();
     }
