+++ src/interfaces/implementations/key-registries/IKeyRegistry.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.0;
 import {IOzEIP712} from "../../base/common/IOzEIP712.sol";
 
 interface IKeyRegistry {
-    function getOperatorsLengthAt(uint48 timestamp, bytes memory hint) external view returns (uint256);
+    function getOperatorsLengthAt(uint48 timestamp, bytes memory hint) external view returns (uint208);
 
-    function getOperatorsLength() external view returns (uint256);
+    function getOperatorsLength() external view returns (uint208);
 
     function getOperatorsAt(uint48 timestamp, bytes[] memory hints) external view returns (address[] memory);
